first picture approaching

+++ shaders/2dMultiTexture.vert
@@ -2,7 +2,7 @@
 
 layout(location = 0) in vec2 position;
 layout(location = 1) in vec2 TexCoord;
-layout(location = 2) in int tex;
+layout(location = 2) in float tex;
 
 out vec2 texCoord;
 out float TexIndex;
@@ -15,5 +15,5 @@ void main()
 {
     gl_Position = projection * view * model * vec4(position, 0.0, 1.0);
     texCoord = TexCoord;
-    TexIndex = tex;
+    TexIndex = float(tex);
 }
